+++ SharkryEngine/SharkryEngine/Engine/Source/Games/LastWar/Characters/PlayerCharacter.cpp
@@ -22,6 +22,7 @@
 #include "Components/SkeletalMeshComponent.h"
 #include "AssetImporter/FBX/FLoaderFBX.h"
 #include "Animation/AnimInstances/AnimSingleNodeInstance.h"
+#include "AssetImporter/FBX/FBXManager.h"
 
 APlayerCharacter::APlayerCharacter()
 {
