+++ SharkryEngine/SharkryEngine/Engine/Source/Runtime/Renderer/SkeletalMeshRenderPass.cpp
@@ -29,6 +29,7 @@
 
 #include "AssetImporter/FBX/FLoaderFBX.h"
 #include "FSkeletalMeshDebugger.h"
+#include "AssetImporter/FBX/FBXStructs.h"
 
 FSkeletalMeshRenderPass::FSkeletalMeshRenderPass()
     : VertexShader(nullptr)
